+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - Added some tests for the management command code.
 - Added a test to prove that the ordering specified on a model's `_meta` is used for pagination, when no custom order has been specified on the query set.
 - Added a `@task_or_admin_only` decorator to `djangae.environment` to allow restricting views to tasks (including crons) or admins of the application.
+- Protected the `cleanup_locks` view in djangae.contrib.locking with the `@task_or_admin_only` decorator.
 
 ### Bug fixes: